Preferences saving implemented

+++ src/main/java/pl/agh/enrollme/service/PreferencesManagementService.java
@@ -158,14 +158,26 @@ public class PreferencesManagementService implements IPreferencesManagementServi
 
         //Persisting points
         for (StudentPointsPerTerm tp : termPoints) {
-            final StudentPointsPerTerm termPoint = pointsDAO.getByPK(tp.getId());
+            StudentPointsPerTerm termPoint;
+            Term term;
+
+            if (tp.getId() != null) {
+                termPoint = pointsDAO.getByPK(tp.getId());
+                term = termPoint.getTerm();
+            } else {
+                termPoint = null;
+            }
 
             //termPoint is not present in the database
-            if (termPoint == null || termPoint.getPoints() != 0) {
-                pointsDAO.add(termPoint);
-                LOGGER.debug("Term points: " + termPoint + " added to the datebase");
-                addedCount++;
-            } else { //is present in the database
+            if (termPoint == null) {
+                if (tp.getPoints() != 0) {
+                    pointsDAO.add(tp);
+                    LOGGER.debug("Term points: " + termPoint + " added to the datebase");
+                    addedCount++;
+                } else {
+                    LOGGER.debug("Term points: " + termPoint + " not present in the datebase, but zero-point");
+                }
+            } else if (!termPoint.getAssigned() && !termPoint.getTerm().getCertain()) { //is present in the database and isn't assigned yet and corresponding term isn't certain
                 if (termPoint.getPoints() == 0) {
                     pointsDAO.remove(termPoint);
                     LOGGER.debug("Term points: " + termPoint + " removed from the datebase");
@@ -179,7 +191,22 @@ public class PreferencesManagementService implements IPreferencesManagementServi
         }
         LOGGER.debug("Points persisted");
 
-        //TODO: Record that the user successfully enrolled and can be exported...
+        final Person person = personService.getCurrentUser();
+        LOGGER.debug("Current person retrieved");
+
+        final List<Subject> subjects = person.getSubjects();
+        final List<Subject> subjectsSaved = person.getSubjectsSaved();
+
+        for (Subject subject : subjects) {
+            if (subject.getHasInteractive() && !subjectsSaved.contains(subject)) {
+                subjectsSaved.add(subject);
+                LOGGER.debug("Subject: " + subject + " added to persons: " + person + " saved subjects list");
+            }
+        }
+        LOGGER.debug("Saved subjects updated");
+
+        personDAO.update(person);
+        LOGGER.debug("Person: " + person + " updated");
 
         FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO, "Choice updated", "Changes successfully saved. "
             + addedCount + " term points have been added, " + updatedCount + " updated, " + removedCount + " removed from the datebase");
@@ -188,7 +215,9 @@ public class PreferencesManagementService implements IPreferencesManagementServi
         LOGGER.debug("Saving finished");
     }
 
-    //Validates if minimum point usage has been reached for all subjects
+    /**
+     * Validates if minimum point usage has been reached for all subjects
+     */
     private boolean validateMinimumReached(Map<Integer, Integer> pointsMap, EnrollConfiguration enrollConfiguration) {
         final Integer minimumPointsPerSubject = enrollConfiguration.getMinimumPointsPerSubject();
         for (Integer used : pointsMap.values()) {
@@ -201,7 +230,9 @@ public class PreferencesManagementService implements IPreferencesManagementServi
         return true;
     }
 
-    //Creates missing SPPTs
+    /**
+     * Creates missing SPPTs
+     */
     private void createMissingSPPT(List<Term> terms, List<StudentPointsPerTerm> points, Person person) {
         Map<Term, Boolean> termsPresent = new HashMap<Term, Boolean>();
